requirements/ci.txt
# Dependencies used by CI on github.com/plotly/dash
black==22.3.0
dash-flow-example==0.0.5
dash-dangerously-set-inner-html
flake8==7.0.0
flaky==3.8.1
flask-talisman==1.0.0
mimesis<=11.1.0
mock==4.0.3
numpy<=1.26.3
orjson==3.10.3
openpyxl
pandas>=1.4.0
pyarrow
pylint==3.0.3
pytest-mock
pytest-sugar==0.9.6
pyzmq==25.1.2
xlrd>=2.0.1
pytest-rerunfailures
jupyterlab<4.0.0
